Fix permission problems + naming in .smw.json directory
Fixes issue #44 "Ensure .smw.json has a home"

The permission problems stem from the fact that a local userX can run ad-hoc shell scripts that will want to read/write; a deploy will read-write as user meza-ansible and of course the web server process will read-write as apache. So all these users and groups can be handled with meza-ansible:wheel

The directory naming problem stems from the choice to use the wgDBname variable instead of the "wiki id". Wiki ID is a better choice for **wiki** related config data directory; and it's also easier to use over db name because the latter is a composed value, not a simple string.

This work  was performed for NASA GRC-ATF by WikiWorks per NASA Contract  NNC15BA02B.