Fix first-time deploy errors related to SMW rebuild data
Add the special 'never' tag to the task in the MediaWiki role.
With the 'never' tag, this task will ONLY be run if you specifically ask
for it with the --tags=smw-data option (or --tags=never option).

This changes behavior of Meza so that it does not rebuild SMW data
on each and every deploy. To specifically request a SMW rebuild, simply
ask for it by name with `sudo meza deploy monolith --tags smw-data`

Fixes Issue #80

This work was performed for NASA GRC-ATF by WikiWorks per NASA Contract NNC15BA02B.