Add a couple of 'ToDo' comments and reformat the smw index shell cmd
The shell command was not working due to block folding the
list_of_wikis into a single string so the cmd was reformatted to a
single line (which is too long for yamllint compliance). But, at least
it nominally 'works'. A comment was added to address the issue that
this maintenance should not even exist in this role, but should be refactored. See issue #117